extract FactoryCompiler
